Refactored many code

+++ TheBank/Models/Account.cs
@@ -3,10 +3,10 @@
     abstract public class Account
     {
         public string Name { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
         public int AccountNumber { get; set; }
         public string AccountType { get; set; }
 
-        public abstract decimal ChargeInterest();
+        public abstract double ChargeInterest();
     }
 }
