+++ TheBank/Models/SavingsAccount.cs
@@ -9,19 +9,19 @@
             AccountType = "Opsparingskonto";
         }
 
-        public override decimal ChargeInterest()
+        public override double ChargeInterest()
         {
             if (Balance > 100000)
             {
-                Balance *= 1.03m;
+                Balance *= 1.03;
             }
             else if (Balance >= 50000)
             {
-                Balance *= 1.02m;
+                Balance *= 1.02;
             }
             else if (Balance < 50000)
             {
-                Balance *= 1.01m;
+                Balance *= 1.01;
             }
             return Balance;
         }
