+++ TheBank/Repository/IBank.cs
@@ -5,10 +5,10 @@ namespace TheBank.Repository
     public interface IBank
     {
         Account? CreateAccount(string name, ConsoleKey accType);
-        decimal? Deposit(int accountNumber, decimal amount);
-        decimal? Withdraw(int accountNumber, decimal amount);
-        decimal? Balance(int accountNumber);
-        decimal BankBalance();
+        double? Deposit(int accountNumber, double amount);
+        double? Withdraw(int accountNumber, double amount);
+        double? Balance(int accountNumber);
+        double BankBalance();
         void ChargeInterest();
         List<AccountListItem> GetAccountList();
     }
